renamed test_ExchangeRateManager to test_exchange_rate_manager
