+++ lbry/tests/unit/lbrynet_daemon/test_exchange_rate_manager.py
